+++ geo/src/algorithm/indexed/mod.rs
@@ -1,0 +1,3 @@
+/// Geometries that are backed by a spatial index for faster bulk operations
+pub mod prepared_geometry;
+pub use prepared_geometry::PreparedGeometry;